requirements.txt
streamlitgoogle-generativeaiPillow